Refactored to extract strategies
